pyproject.toml
[project]
name = "comic-downloader"
version = "0.1.0"
description = "Script to download comics from various sources."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "curl-cffi==0.15.0",
    "heapdict>=1.0.1",
    "openpyxl==3.1.5",
    "pandas==3.0.2",
    "pillow==12.2.0",
    "playwright==1.58.0",
    "pydantic==2.12.5",
    "pydantic-settings==2.13.1",
    "selectolax==0.4.7",
]
